# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blockchain Node API - Laravel",
        "description": "API REST para nodo blockchain distribuido de grados académicos.",
        "contact": {
            "email": "admin@blockchain-node.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8004",
            "description": "Nodo Laravel - Puerto 8004"
        }
    ],
    "paths": {
        "/chain": {
            "get": {
                "tags": [
                    "Blockchain"
                ],
                "summary": "Obtener la cadena completa de bloques",
                "operationId": "c928152df4768b6704537de8b32373df",
                "responses": {
                    "200": {
                        "description": "Cadena obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "chain": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "length": {
                                            "type": "integer",
                                            "example": 2
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/mine": {
            "post": {
                "tags": [
                    "Blockchain"
                ],
                "summary": "Minar transacciones pendientes",
                "operationId": "e968987e86d7dab536895ebae181dffb",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "firmado_por": {
                                        "type": "string",
                                        "example": "nodo-laravel-8004"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Bloques minados correctamente"
                    },
                    "400": {
                        "description": "No hay transacciones pendientes"
                    }
                }
            }
        },
        "/api/blocks/receive": {
            "post": {
                "tags": [
                    "Blockchain"
                ],
                "summary": "Recibir un bloque de otro nodo",
                "operationId": "8694001819fdffff46f3b35d447b79f7",
                "responses": {
                    "201": {
                        "description": "Bloque aceptado"
                    },
                    "400": {
                        "description": "Bloque inválido"
                    }
                }
            }
        },
        "/api/genesis": {
            "post": {
                "tags": [
                    "Blockchain"
                ],
                "summary": "Crear el bloque génesis",
                "operationId": "8b151eeda72866ff2cb4899f16c68447",
                "responses": {
                    "201": {
                        "description": "Bloque génesis creado"
                    },
                    "400": {
                        "description": "La cadena ya tiene bloques"
                    }
                }
            }
        },
        "/api/health": {
            "get": {
                "tags": [
                    "Blockchain"
                ],
                "summary": "Estado del nodo",
                "operationId": "1d94265a9d4dffe77b843709306bbf12",
                "responses": {
                    "200": {
                        "description": "Estado del nodo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "ok"
                                        },
                                        "node_id": {
                                            "type": "string",
                                            "example": "nodo-laravel-8004"
                                        },
                                        "pendientes": {
                                            "type": "integer",
                                            "example": 0
                                        },
                                        "bloques": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/nodes/register": {
            "post": {
                "tags": [
                    "Nodos"
                ],
                "summary": "Registrar un nodo en la red",
                "operationId": "ee8c3e97653d07dba9771925290b2191",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "url"
                                ],
                                "properties": {
                                    "url": {
                                        "type": "string",
                                        "example": "http://localhost:8001"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Nodo registrado correctamente"
                    },
                    "422": {
                        "description": "URL inválida"
                    }
                }
            }
        },
        "/nodes": {
            "get": {
                "tags": [
                    "Nodos"
                ],
                "summary": "Listar nodos registrados",
                "operationId": "5d2024693a7063efd7dfdc2d8f68fd86",
                "responses": {
                    "200": {
                        "description": "Lista de nodos"
                    }
                }
            }
        },
        "/nodes/resolve": {
            "get": {
                "tags": [
                    "Nodos"
                ],
                "summary": "Resolver conflictos por consenso (longest chain)",
                "operationId": "17f47dbe75f576e4b491b05cb2121abc",
                "responses": {
                    "200": {
                        "description": "Consenso ejecutado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensaje": {
                                            "type": "string",
                                            "example": "La cadena local ya es la más larga"
                                        },
                                        "longitud": {
                                            "type": "integer",
                                            "example": 5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/transactions": {
            "post": {
                "tags": [
                    "Transacciones"
                ],
                "summary": "Crear una nueva transacción",
                "description": "Recibe los datos de un grado académico, los guarda como transacción pendiente y los propaga a todos los nodos registrados en la red.",
                "operationId": "238affdbe033b73010cbace3ad94520e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "persona_id",
                                    "institucion_id",
                                    "programa_id",
                                    "titulo_obtenido",
                                    "fecha_fin"
                                ],
                                "properties": {
                                    "persona_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "286cd43f-f789-4903-9ae6-de288bd81e25"
                                    },
                                    "institucion_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "8a488bb8-f8d8-48da-a88e-2abd852aea84"
                                    },
                                    "programa_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "6df0a016-a1d7-4ee4-ac92-9af454b9454f"
                                    },
                                    "titulo_obtenido": {
                                        "type": "string",
                                        "example": "Ingeniero en Sistemas Computacionales"
                                    },
                                    "fecha_inicio": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2019-08-01"
                                    },
                                    "fecha_fin": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-06-15"
                                    },
                                    "numero_cedula": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "titulo_tesis": {
                                        "type": "string",
                                        "example": "Implementación de redes neuronales"
                                    },
                                    "menciones": {
                                        "type": "string",
                                        "example": "Cum Laude"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transacción creada y propagada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensaje": {
                                            "type": "string",
                                            "example": "Transacción agregada y propagada"
                                        },
                                        "transaccion": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/transactions/pending": {
            "get": {
                "tags": [
                    "Transacciones"
                ],
                "summary": "Listar transacciones pendientes",
                "description": "Retorna todas las transacciones que aún no han sido incluidas en un bloque minado.",
                "operationId": "09b53d9ba377035d08860b08e5992efd",
                "responses": {
                    "200": {
                        "description": "Lista de transacciones pendientes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "transacciones_pendientes": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Blockchain",
            "description": "Cadena de bloques y minado"
        },
        {
            "name": "Transacciones",
            "description": "Gestión de transacciones pendientes"
        },
        {
            "name": "Nodos",
            "description": "Registro y sincronización de nodos"
        }
    ]
}